Warning fixes : uint64 to size_t conversions.
git-svn-id: http://svn.ta3d.org/yuni/trunk@2241 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/stream.cpp
@@ -194,7 +194,7 @@ namespace File
 	{
 		//TODO We should not read with block too large (especially on 32bits)
 		// systems. It can make the process/thread interruptible  too long
-		return (uint64) ::fread(buffer, 1, size, pFd);
+		return (uint64) ::fread(buffer, 1, (size_t)size, pFd);
 	}
